requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==20.1.0
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.1


